Added concave math (not yet in use). Updated UnitMatchers to handle any player's units. Updated PvP game plan to account for FFE builds.

+++ src/Mathematics/Points/Pixel.scala
@@ -48,7 +48,7 @@ case class Pixel(val x:Int, val y:Int) {
     delta.multiply(pixels/distance).add(this)
   }
   private val radiansOverAngle = 2.0 * Math.PI / 256.0
-  def radiate(angle:Int, pixels:Double):Pixel = {
+  def radiate(angle: Double, pixels: Double):Pixel = {
     // According to JohnJ, Brood War understands 256 angles
     // The BWAPI interface reduces this to a double (radians) with the cartesian origin (pointing right)
     // We'll use 256 (to match the engine behavior) and the BWAPI origin
